+++ bedrock-models.js
@@ -7,11 +7,68 @@
 
 export const bedrock_models = [
     {
+        // ================================
+        // == Claude 3.7 Sonnet Thinking ==
+        // ================================
+        "modelName":                     "Claude-3-7-Sonnet-Thinking",
+        // "modelId":                       "anthropic.claude-3-7-sonnet-20250219-v1:0",
+        "modelId":                       "us.anthropic.claude-3-7-sonnet-20250219-v1:0",
+        "vision":                        true,
+        "messages_api":                  true,
+        "system_as_separate_field":      true,
+        "display_role_names":            true,
+        "max_tokens_param_name":         "max_tokens",
+        "max_supported_response_tokens": 131072,
+        "response_chunk_element":        "delta.text",
+        "response_nonchunk_element":     "content[0].text",
+        "thinking_response_chunk_element": "delta.thinking",
+        "thinking_response_nonchunk_element": "content[0].thinking",
+        "special_request_schema": {
+            "anthropic_version": "bedrock-2023-05-31",
+            "anthropic_beta": ["output-128k-2025-02-19"],
+            "thinking": {
+                "type": "enabled",
+                "budget_tokens": 16000
+            },
+        },
+        "image_support": {
+            "max_image_size": 20971520, // 20MB
+            "supported_formats": ["jpeg", "png", "gif", "webp"],
+            "max_images_per_request": 10
+        }
+    },
+    {
+        // =======================
+        // == Claude 3.7 Sonnet ==
+        // =======================
+        "modelName":                     "Claude-3-7-Sonnet",
+        // "modelId":                       "anthropic.claude-3-7-sonnet-20250219-v1:0",
+        "modelId":                       "us.anthropic.claude-3-7-sonnet-20250219-v1:0",
+        "vision":                        true,
+        "messages_api":                  true,
+        "system_as_separate_field":      true,
+        "display_role_names":            true,
+        "max_tokens_param_name":         "max_tokens",
+        "max_supported_response_tokens": 131072,
+        "response_chunk_element":        "delta.text",
+        "response_nonchunk_element":     "content[0].text",
+        "special_request_schema": {
+            "anthropic_version": "bedrock-2023-05-31",
+            "anthropic_beta": ["output-128k-2025-02-19"]
+        },
+        "image_support": {
+            "max_image_size": 20971520, // 20MB
+            "supported_formats": ["jpeg", "png", "gif", "webp"],
+            "max_images_per_request": 10
+        }
+    },
+    {
         // ==========================
         // == Claude 3.5 Sonnet v2 ==
         // ==========================
         "modelName":                     "Claude-3-5-Sonnet-v2",
         "modelId":                       "anthropic.claude-3-5-sonnet-20241022-v2:0",
+        "vision":                        true,
         "messages_api":                  true,
         "system_as_separate_field":      true,
         "display_role_names":            true,
@@ -21,6 +78,11 @@ export const bedrock_models = [
         "response_nonchunk_element":     "content[0].text",
         "special_request_schema": {
             "anthropic_version": "bedrock-2023-05-31"
+        },
+        "image_support": {
+            "max_image_size": 20971520, // 20MB
+            "supported_formats": ["jpeg", "png", "gif", "webp"],
+            "max_images_per_request": 10
         }
     },
     {
@@ -29,6 +91,7 @@ export const bedrock_models = [
         // =======================
         "modelName":                     "Claude-3-5-Sonnet",
         "modelId":                       "anthropic.claude-3-5-sonnet-20240620-v1:0",
+        "vision":                        true,
         "messages_api":                  true,
         "system_as_separate_field":      true,
         "display_role_names":            true,
@@ -38,6 +101,11 @@ export const bedrock_models = [
         "response_nonchunk_element":     "content[0].text",
         "special_request_schema": {
             "anthropic_version": "bedrock-2023-05-31"
+        },
+        "image_support": {
+            "max_image_size": 20971520, // 20MB
+            "supported_formats": ["jpeg", "png", "gif", "webp"],
+            "max_images_per_request": 10
         }
     },
     {
@@ -46,6 +114,7 @@ export const bedrock_models = [
         // ======================
         "modelName":                     "Claude-3-5-Haiku",
         "modelId":                       "anthropic.claude-3-5-haiku-20241022-v1:0",
+        "vision":                        false,
         "messages_api":                  true,
         "system_as_separate_field":      true,
         "display_role_names":            true,
@@ -63,6 +132,7 @@ export const bedrock_models = [
         // ====================
         "modelName":                     "Claude-3-Haiku",
         "modelId":                       "anthropic.claude-3-haiku-20240307-v1:0",
+        "vision":                        false,
         "messages_api":                  true,
         "system_as_separate_field":      true,
         "display_role_names":            true,
@@ -72,6 +142,11 @@ export const bedrock_models = [
         "response_nonchunk_element":     "content[0].text",
         "special_request_schema": {
             "anthropic_version": "bedrock-2023-05-31"
+        },
+        "image_support": {
+            "max_image_size": 20971520, // 20MB
+            "supported_formats": ["jpeg", "png", "gif", "webp"],
+            "max_images_per_request": 10
         }
     },
     {
@@ -81,6 +156,7 @@ export const bedrock_models = [
         "modelName":                     "Llama-3-3-70b",
         // "modelId":                       "meta.llama3-3-70b-instruct-v1:0",
         "modelId":                       "us.meta.llama3-3-70b-instruct-v1:0",
+        "vision":                        false,
         "messages_api":                  false,
         "bos_text":                      "<|begin_of_text|>",
         "role_system_message_prefix":    "",
@@ -99,7 +175,7 @@ export const bedrock_models = [
         "display_role_names":            true,
         "max_tokens_param_name":         "max_gen_len",
         "max_supported_response_tokens": 2048,
-        "response_chunk_element":        "generation",
+        "response_chunk_element":        "generation"
     },
     {
         // ==================
@@ -108,6 +184,7 @@ export const bedrock_models = [
         "modelName":                     "Llama-3-2-1b",
         // "modelId":                       "meta.llama3-2-1b-instruct-v1:0",
         "modelId":                       "us.meta.llama3-2-1b-instruct-v1:0",
+        "vision":                        false,
         "messages_api":                  false,
         "bos_text":                      "<|begin_of_text|>",
         "role_system_message_prefix":    "",
@@ -126,7 +203,7 @@ export const bedrock_models = [
         "display_role_names":            true,
         "max_tokens_param_name":         "max_gen_len",
         "max_supported_response_tokens": 2048,
-        "response_chunk_element":        "generation",
+        "response_chunk_element":        "generation"
     },
     {
         // ==================
@@ -135,6 +212,7 @@ export const bedrock_models = [
         "modelName":                     "Llama-3-2-3b",
         // "modelId":                       "meta.llama3-2-3b-instruct-v1:0",
         "modelId":                       "us.meta.llama3-2-3b-instruct-v1:0",
+        "vision":                        false,
         "messages_api":                  false,
         "bos_text":                      "<|begin_of_text|>",
         "role_system_message_prefix":    "",
@@ -153,7 +231,7 @@ export const bedrock_models = [
         "display_role_names":            true,
         "max_tokens_param_name":         "max_gen_len",
         "max_supported_response_tokens": 2048,
-        "response_chunk_element":        "generation",
+        "response_chunk_element":        "generation"
     },
     {
         // ===================
@@ -162,6 +240,7 @@ export const bedrock_models = [
         "modelName":                     "Llama-3-2-11b",
         // "modelId":                       "meta.llama3-2-11b-instruct-v1:0",
         "modelId":                       "us.meta.llama3-2-11b-instruct-v1:0",
+        "vision":                        false,
         "messages_api":                  false,
         "bos_text":                      "<|begin_of_text|>",
         "role_system_message_prefix":    "",
@@ -180,7 +259,7 @@ export const bedrock_models = [
         "display_role_names":            true,
         "max_tokens_param_name":         "max_gen_len",
         "max_supported_response_tokens": 2048,
-        "response_chunk_element":        "generation",
+        "response_chunk_element":        "generation"
     },
     {
         // ===================
@@ -189,6 +268,7 @@ export const bedrock_models = [
         "modelName":                     "Llama-3-2-90b",
         // "modelId":                       "meta.llama3-2-90b-instruct-v1:0",
         "modelId":                       "us.meta.llama3-2-90b-instruct-v1:0",
+        "vision":                        false,
         "messages_api":                  false,
         "bos_text":                      "<|begin_of_text|>",
         "role_system_message_prefix":    "",
@@ -207,7 +287,7 @@ export const bedrock_models = [
         "display_role_names":            true,
         "max_tokens_param_name":         "max_gen_len",
         "max_supported_response_tokens": 2048,
-        "response_chunk_element":        "generation",
+        "response_chunk_element":        "generation"
     },
     {
         // ==================
@@ -215,6 +295,7 @@ export const bedrock_models = [
         // ==================
         "modelName":                     "Llama-3-1-8b",
         "modelId":                       "meta.llama3-1-8b-instruct-v1:0",
+        "vision":                        false,
         "messages_api":                  false,
         "bos_text":                      "<|begin_of_text|>",
         "role_system_message_prefix":    "",
@@ -233,7 +314,7 @@ export const bedrock_models = [
         "display_role_names":            true,
         "max_tokens_param_name":         "max_gen_len",
         "max_supported_response_tokens": 2048,
-        "response_chunk_element":        "generation",
+        "response_chunk_element":        "generation"
     },
     {
         // ===================
@@ -241,6 +322,7 @@ export const bedrock_models = [
         // ===================
         "modelName":                     "Llama-3-1-70b",
         "modelId":                       "meta.llama3-1-70b-instruct-v1:0",
+        "vision":                        false,
         "messages_api":                  false,
         "bos_text":                      "<|begin_of_text|>",
         "role_system_message_prefix":    "",
@@ -259,7 +341,7 @@ export const bedrock_models = [
         "display_role_names":            true,
         "max_tokens_param_name":         "max_gen_len",
         "max_supported_response_tokens": 2048,
-        "response_chunk_element":        "generation",
+        "response_chunk_element":        "generation"
     },
     {
         // ====================
@@ -267,6 +349,7 @@ export const bedrock_models = [
         // ====================
         "modelName":                     "Llama-3-1-405b",
         "modelId":                       "meta.llama3-1-405b-instruct-v1:0",
+        "vision":                        false,
         "messages_api":                  false,
         "bos_text":                      "<|begin_of_text|>",
         "role_system_message_prefix":    "",
@@ -285,7 +368,7 @@ export const bedrock_models = [
         "display_role_names":            true,
         "max_tokens_param_name":         "max_gen_len",
         "max_supported_response_tokens": 2048,
-        "response_chunk_element":        "generation",
+        "response_chunk_element":        "generation"
     },
     {
         // ================
@@ -293,6 +376,7 @@ export const bedrock_models = [
         // ================
         "modelName":                     "Llama-3-8b",
         "modelId":                       "meta.llama3-8b-instruct-v1:0",
+        "vision":                        false,
         "messages_api":                  false,
         "bos_text":                      "<|begin_of_text|>",
         "role_system_message_prefix":    "",
@@ -311,7 +395,7 @@ export const bedrock_models = [
         "display_role_names":            true,
         "max_tokens_param_name":         "max_gen_len",
         "max_supported_response_tokens": 2048,
-        "response_chunk_element":        "generation",
+        "response_chunk_element":        "generation"
     },
     {
         // =================
@@ -319,6 +403,7 @@ export const bedrock_models = [
         // =================
         "modelName":                     "Llama-3-70b",
         "modelId":                       "meta.llama3-70b-instruct-v1:0",
+        "vision":                        false,
         "messages_api":                  false,
         "bos_text":                      "<|begin_of_text|>",
         "role_system_message_prefix":    "",
@@ -337,7 +422,7 @@ export const bedrock_models = [
         "display_role_names":            true,
         "max_tokens_param_name":         "max_gen_len",
         "max_supported_response_tokens": 2048,
-        "response_chunk_element":        "generation",
+        "response_chunk_element":        "generation"
     },
     {
         // ================
@@ -345,6 +430,7 @@ export const bedrock_models = [
         // ================
         "modelName":                     "Mistral-7b",
         "modelId":                       "mistral.mistral-7b-instruct-v0:2",
+        "vision":                        false,
         "messages_api":                  false,
         "bos_text":                      "<s>",
         "role_system_message_prefix":    "",
@@ -363,7 +449,7 @@ export const bedrock_models = [
         "display_role_names":            false,
         "max_tokens_param_name":         "max_tokens",
         "max_supported_response_tokens": 8192,
-        "response_chunk_element":        "outputs[0].text",
+        "response_chunk_element":        "outputs[0].text"
     },
     {
         // ==================
@@ -371,6 +457,7 @@ export const bedrock_models = [
         // ==================
         "modelName":                     "Mixtral-8x7b",
         "modelId":                       "mistral.mixtral-8x7b-instruct-v0:1",
+        "vision":                        false,
         "messages_api":                  false,
         "bos_text":                      "<s>",
         "role_system_message_prefix":    "",
@@ -389,7 +476,7 @@ export const bedrock_models = [
         "display_role_names":            false,
         "max_tokens_param_name":         "max_tokens",
         "max_supported_response_tokens": 4096,
-        "response_chunk_element":        "outputs[0].text",
+        "response_chunk_element":        "outputs[0].text"
     },
     {
         // ===================
@@ -397,6 +484,7 @@ export const bedrock_models = [
         // ===================
         "modelName":                     "Mistral-Large",
         "modelId":                       "mistral.mistral-large-2402-v1:0",
+        "vision":                        false,
         "messages_api":                  false,
         "bos_text":                      "<s>",
         "role_system_message_prefix":    "",
@@ -415,6 +503,6 @@ export const bedrock_models = [
         "display_role_names":            false,
         "max_tokens_param_name":         "max_tokens",
         "max_supported_response_tokens": 8192,
-        "response_chunk_element":        "outputs[0].text",
+        "response_chunk_element":        "outputs[0].text"
     },
 ];